refactor: isolate InputGroupContext file
